Moved Sae tests class into 'Drupal\Tests\metastore\Unit\Sae'
Helper classes are autoloadable now.
