--- modules/metastore/tests/src/Sae/Memory.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Drupal\Tests\metastore\Sae;
-
-use Contracts\BulkRetrieverInterface;
-
-class Memory extends UnsupportedMemory implements BulkRetrieverInterface
-{
-
-    private $storage = [];
-
-    public function retrieveAll(): array
-    {
-        return $this->storage;
-    }
-}
